requirements.txt
requests==2.32.3
pandas==2.2.2
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
scikit-learn==1.5.1
python-dotenv==1.0.1
joblib==1.4.2